requirements.txt
pandas
requests
boto3
python-dotenv